+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+- Fixed a bug where the default entry type could be used for the Child Me! dropdown menu, even if it has been filtered out via an `EVENT_DEFINE_ENTRY_TYPES` event handler  
+
 ## 1.3.2 - 2023-02-22
 ### Fixed  
 - Fixed some styling issues for the entry type menus on Craft 4.3